Uploading Directory
UDWidget version god knows what now. I'm just going to call it v1.4.1
seems that seems closest to accurate. This commit contains updated code
to the Zombienames.js that currently broke with a style update to
UrbanDead and should now be functional. If you know how to replace the
extension repository you can do it there.